+++ schema.graphql
@@ -5473,14 +5473,14 @@ type Mutation {
   ): ApiKey
 
   """
-  Sync crm integration invoice
+  Sync hubspot integration invoice
   """
-  syncCrmIntegrationInvoice(
+  syncHubspotIntegrationInvoice(
     """
-    Parameters for SyncCrmIntegrationInvoice
+    Parameters for SyncHubspotInvoice
     """
-    input: SyncCrmIntegrationInvoiceInput!
-  ): SyncCrmIntegrationInvoicePayload
+    input: SyncHubspotIntegrationInvoiceInput!
+  ): SyncHubspotInvoicePayload
 
   """
   Sync integration credit note
@@ -5503,7 +5503,7 @@ type Mutation {
   ): SyncIntegrationInvoicePayload
 
   """
-  Sync Salesforce invoice
+  Sync Salesforce integration invoice
   """
   syncSalesforceInvoice(
     """
@@ -7066,9 +7066,9 @@ type SubsidiaryCollection {
 }
 
 """
-Autogenerated input type of SyncCrmIntegrationInvoice
+Autogenerated input type of SyncHubspotInvoice
 """
-input SyncCrmIntegrationInvoiceInput {
+input SyncHubspotIntegrationInvoiceInput {
   """
   A unique identifier for the client performing the mutation.
   """
@@ -7077,9 +7077,9 @@ input SyncCrmIntegrationInvoiceInput {
 }
 
 """
-Autogenerated return type of SyncCrmIntegrationInvoice.
+Autogenerated return type of SyncHubspotInvoice.
 """
-type SyncCrmIntegrationInvoicePayload {
+type SyncHubspotInvoicePayload {
   """
   A unique identifier for the client performing the mutation.
   """
