+++ schema.json
@@ -26619,11 +26619,11 @@
               ]
             },
             {
-              "name": "syncCrmIntegrationInvoice",
-              "description": "Sync crm integration invoice",
+              "name": "syncHubspotIntegrationInvoice",
+              "description": "Sync hubspot integration invoice",
               "type": {
                 "kind": "OBJECT",
-                "name": "SyncCrmIntegrationInvoicePayload",
+                "name": "SyncHubspotInvoicePayload",
                 "ofType": null
               },
               "isDeprecated": false,
@@ -26631,13 +26631,13 @@
               "args": [
                 {
                   "name": "input",
-                  "description": "Parameters for SyncCrmIntegrationInvoice",
+                  "description": "Parameters for SyncHubspotInvoice",
                   "type": {
                     "kind": "NON_NULL",
                     "name": null,
                     "ofType": {
                       "kind": "INPUT_OBJECT",
-                      "name": "SyncCrmIntegrationInvoiceInput",
+                      "name": "SyncHubspotIntegrationInvoiceInput",
                       "ofType": null
                     }
                   },
@@ -26707,7 +26707,7 @@
             },
             {
               "name": "syncSalesforceInvoice",
-              "description": "Sync Salesforce invoice",
+              "description": "Sync Salesforce integration invoice",
               "type": {
                 "kind": "OBJECT",
                 "name": "SyncSalesforceInvoicePayload",
@@ -37160,8 +37160,8 @@
         },
         {
           "kind": "INPUT_OBJECT",
-          "name": "SyncCrmIntegrationInvoiceInput",
-          "description": "Autogenerated input type of SyncCrmIntegrationInvoice",
+          "name": "SyncHubspotIntegrationInvoiceInput",
+          "description": "Autogenerated input type of SyncHubspotInvoice",
           "interfaces": null,
           "possibleTypes": null,
           "fields": null,
@@ -37199,8 +37199,8 @@
         },
         {
           "kind": "OBJECT",
-          "name": "SyncCrmIntegrationInvoicePayload",
-          "description": "Autogenerated return type of SyncCrmIntegrationInvoice.",
+          "name": "SyncHubspotInvoicePayload",
+          "description": "Autogenerated return type of SyncHubspotInvoice.",
           "interfaces": [
 
           ],
